Move helper to api utils
